==> 02_itn-eval-metrics_plots_metricminer.Rmd <==
---
title: "ITN Eval Metrics"
author: "Candace Savonen"
date: "2024-01-05"
output:
  html_document:
    code_folding: hide
    toc: true
    toc_depth: 2
    toc_float: true
  word_document: default
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Set Up
```{r}
library("ggplot2")
library("magrittr")
library("tidyverse")
library("ggrepel")
library("patchwork")

cbPalette <- c("#E69F02", "#56B4E9", "#009E73", "#008080") # Everyone, Leadership, new to data science, software developers, 
```


## Load data
```{r}
ga_dims <- readRDS(file.path("data","itcr_ga_dims_data.RDS"))
ga_link_clicks <- readRDS(file.path("data","itcr_ga_link_click_data.RDS"))
```

## Web traffic

### Load Web Traffic data

```{r}
itcr_course_data <- read_tsv(file.path("data", "itcr_course_metrics.tsv")) %>% 
  mutate(target_audience = replace_na(target_audience, "Everyone"))

ga_metrics <- readRDS(file.path("data","itcr_ga_metric_data.RDS"))
itcr_website_metrics <- read_tsv(file.path("data", "itcr_website_metrics.tsv"))
web_traffic_overtime <- read_tsv(file.path("data", "web_traffic_overtime.tsv"))
```

### Unique visitors to websites

READY
```{r}
ggplot(itcr_course_data, aes(x = reorder(website, -totalUsers), 
                             y = totalUsers, 
                             fill = target_audience)) +
  geom_bar(stat = "identity") +
  geom_text(aes(label = totalUsers), size = 3, vjust = - 1) +
  theme_classic() +
  theme(axis.text.x = element_text(angle = 45, hjust=1)) +
  labs(x = NULL,
       y = "Total # of Visitors",
       fill = "Target Audience",
       title = "Visitor Distribution across Educational Resources",
       caption = "The graph displays the number of visitors for different educational platforms, categorized by the intended audience: everyone, leadership, newcomers to data, and software developers.") +
  ylim(c(0, 6000)) + 
  scale_fill_manual(values=cbPalette)
```

### Table of user engagement

```{r}
user_totals <- ga_metrics %>% 
  janitor::clean_names() %>% 
  select(website, active_users, average_session_duration) %>% 
  knitr::kable(digits=2, long_table = TRUE,padding = 2) # %>%

user_totals
```

```{r}
user_engagement <- ga_metrics %>% 
  janitor::clean_names() %>% 
  select(website, screen_page_views_per_user, sessions, screen_page_views, engagement_rate) %>% 
  knitr::kable(digits=2, long_table = TRUE, padding = 2) # %>%

user_engagement
```


READY
```{r}
itcr_course_data %>% janitor::clean_names() %>%
  select(website, screen_page_views_per_user, average_session_duration, event_count_per_user, engagement_rate, target_audience) %>%
  tidyr::pivot_longer(!c(website, target_audience), 
                      names_to = "metric_name", 
                      values_to = "value") %>% 
  filter(!(website %in% c("ITN Website", "OTTR website", "metricminer.org"))) %>%
  ggplot(aes(x = website, y = value, fill = target_audience)) +
  geom_bar(position = "dodge", stat = "identity") +
  theme_minimal() +
  labs(x = NULL,
       y = NULL,
       fill = "Target Audience") +
  facet_wrap(~metric_name, scales = "free_y",
             labeller = labeller(metric_name = c(screen_page_views_per_user = "Screen Page Views per User",
                                                 average_session_duration = "Average Session Duration",
                                                 event_count_per_user = "Event Count per User",
                                                 engagement_rate = "Engagement Rate"))) + 
  scale_fill_manual(values=cbPalette) +
  scale_x_discrete(limits = c("Leadership in Cancer Informatics", "NIH Data Sharing", "Ethical Data Handling", "Overleaf and Latex for Scientific Articles", "AI for Decision Makers",
                              "Reproducibility in Cancer Informatics", "Choosing Genomics Tools", "Computing for Cancer Informatics",
                              "Documentation and Usability", "Advanced Reproducibility", "AI for Efficient Programming", "GitHub Automation for Scientists")) +
  theme(axis.text.x=element_text(angle=60, hjust=1), 
        strip.text.x = element_text(size = 8), 
        plot.margin = unit(c(1.5,.5,.5,1.5), "cm"))

ggsave(file.path("plots", "itn_engagement_stats_reordered.png"), width = 8, height = 6)
```

### ITN Course and Website Traffic Overtime 

```{r}
# TODO: 'web_traffic_overtime' not found

traffic_plot <- web_traffic_overtime %>% 
  filter(website %in% c("ITN Website", "OTTR website", "metricminer.org")) %>%
  group_by(month_year) %>% 
  count() %>% 
  ggplot(aes(y = n, x = month_year)) + 
  geom_bar(stat = "identity", fill = "pink") + 
  scale_x_date(date_labels = "%b %Y") + 
  ylab("Unique ITN Online Course Visitors") +
  xlab("") +
  theme_minimal() +
  theme(axis.text.x=element_text(angle=60, hjust=1)) 

ggsave(file.path("plots", "itn_course_traffic_plot.png"), width = 4, height = 3.5)

traffic_plot
```


```{r}
long_df <- itcr_course_data %>% select(c("website", 
                                         "totalUsers",
                                         "coursera_count", 
                                         "leanpub_count", 
                                         "target_audience")) %>%
  tidyr::pivot_longer(!c(website, target_audience),
                      names_to = "modality", 
                      values_to = "learner_count") %>%
  filter(!(website %in% c("ITN Website", "OTTR website", "metricminer.org"))) %>%
  mutate(modality = case_when(
    modality == "leanpub_count" ~ "Total Leanpub Enrollments", 
    modality == "coursera_count" ~ "Total Coursera Enrollments",
    modality == "totalUsers" ~ "Website Learners", 
    TRUE ~ modality
  ))
```


### Total learners from each modality

READY
```{r}
long_df %>% 
  group_by(modality, target_audience) %>% 
  summarize(total_learners = sum(learner_count, na.rm = TRUE)) %>%
  ggplot(aes(x = reorder(modality, -total_learners), y = total_learners, fill = target_audience)) +
  geom_bar(stat = "identity", na.rm = TRUE) +
  theme_classic() +
  theme(axis.text.x = element_text(angle = 45, hjust=1)) +
  labs(x = NULL,
       y = "Visitors/Enrollees",
       fill = "Target Audience") +
  geom_text(aes(label = total_learners), size = 3, vjust = - 1, na.rm = TRUE) + 
  ylim(c(0, 4200)) + 
  facet_wrap(~target_audience) + 
  scale_fill_manual(values=cbPalette)

ggsave(file.path("plots", "itn_total_enrollments.png"), width = 8, height = 3)
```

### Total learners by course

READY
```{r}  
long_df %>% 
  group_by(website, target_audience) %>% 
  summarize(total_learners = sum(learner_count, na.rm = TRUE)) %>%
  ggplot(aes(y = total_learners, x = reorder(website, -total_learners), fill = target_audience)) + 
  geom_bar(stat = "identity") + 
  labs(x = NULL, 
       y = "Total Learners by Course",
       fill = "Target Audience") +
  theme_minimal() +
  theme(axis.text.x=element_text(angle=60, hjust=1), 
        strip.text.x = element_text(size = 8)) + 
  geom_text(aes(label = total_learners), size = 3, vjust = - 1, na.rm = TRUE) +
  ylim(c(0, 1800)) + 
  scale_fill_manual(values=cbPalette)

ggsave(file.path("plots", "itn_total_learners_by_course.png"), width = 10, height = 5)
```

```{r}
long_df %>% 
  group_by(website, target_audience) %>%
  summarize(total_learners = sum(learner_count, na.rm = TRUE)) %>%
  arrange(-total_learners) %>% 
  knitr::kable()
```

### Course traffic by course

```{r}
#publishing dates for each course on the website to add vertical dashes lines to the following plot
publishdates <- itcr_course_data[!(itcr_course_data$website %in% c("ITN Website", "OTTR website", "metricminer.org")),c("website", "website_launch", "coursera_launch", "leanpub_launch")]
publishdates <- publishdates[order(publishdates$website),]
```

```{r}
# TODO: 'web_traffic_overtime' not found
web_traffic_overtime %>%
  group_by(website, month_year, target_audience) %>% 
  count() %>% 
  filter(!(website %in% c("ITN Website", "OTTR website", "metricminer.org"))) %>% 
  ggplot(aes(y = n, x = month_year, fill = target_audience)) + 
  geom_bar(stat = "identity") + 
  scale_x_date(date_labels = "%b %Y") + 
  ylab("Unique ITN Website Learners Overtime") +
  theme_minimal() +
  theme(axis.text.x=element_text(angle=60, hjust=1), 
        strip.text.x = element_text(size = 8)) + 
  facet_wrap(~website) + 
  scale_fill_manual(values=cbPalette) +
  geom_vline(data = publishdates, aes(xintercept = website_launch), linetype='dashed') +
  geom_vline(data = publishdates, aes(xintercept = coursera_launch), linetype='dashed', colour = 'darkgrey') +
  geom_vline(data = publishdates, aes(xintercept = leanpub_launch), linetype='dashed', colour = 'chocolate4')

ggsave(file.path("plots", "itn_course_traffic_plot_by_course.png"), width = 10, height = 5)
```

```{r}
get_plot <- function(course_name, web_traffic_overtime, publishdates, titleOI, useylab = FALSE, usexlab = FALSE, xlimMinVal = min(web_traffic_overtime$month_year), xlimMaxVal = max(web_traffic_overtime$month_year), ylimVal = 400){
  publishdatesSubset <- publishdates[which(publishdates$website == course_name),]
  plotOI <- web_traffic_overtime %>% 
    group_by(website, month_year, target_audience) %>%
    count() %>% #count number of times it is there so it's all the data?
    filter(website == course_name) %>%
    ggplot(aes(y = n, x = month_year, fill = target_audience)) +
    geom_bar(stat = "identity") +
    scale_x_date(date_labels = "%b %Y") +
    xlab("") +
    ylab("") +
    labs(title = titleOI) +
    xlim(xlimMinVal, xlimMaxVal) +
    ylim(0, ylimVal) +
    theme_minimal() +
    theme(axis.text.x=element_text(angle=60, hjust=1),
          strip.text.x = element_text(size=7), axis.title=element_text(size=9), legend.position = "bottom") +
    scale_fill_manual(values=c("Leadership" = "#E69F02", "New to data" = "#56B4E9", "Software developers" = "#009E73")) +
    geom_vline(data = publishdatesSubset, aes(xintercept = website_launch), colour = 'black', linetype='dashed') +
    geom_vline(data = publishdatesSubset, aes(xintercept = coursera_launch), colour='plum4', linetype='dotted', linewidth=1.25) +
    geom_vline(data = publishdatesSubset, aes(xintercept = leanpub_launch), colour='chocolate4', linetype='dashed')
  
  if(useylab){
    plotOI <- plotOI + ylab("Unique ITN Website Learners Overtime")
  }
  
  if(usexlab){
    plotOI <- plotOI + xlab("month_year")
  }
  
  return (plotOI)
}

p00 <- get_plot("Leadership in Cancer Informatics", web_traffic_overtime, publishdates, "Leadership in Cancer\nInformatics") #Leadership in Cancer Informatics

p01 <- get_plot("NIH Data Sharing", web_traffic_overtime, publishdates, "NIH Data Sharing") #NIH Data Sharing

p02 <- get_plot("Ethical Data Handling", web_traffic_overtime, publishdates, "Ethical Data Handling") #Ethical Data Handling

p03 <- get_plot("Overleaf and Latex for Scientific Articles", web_traffic_overtime, publishdates, "Overleaf and Latex for\nScientific Articles") #Overleaf and Latex for Scientific Articles

p04 <- get_plot("AI for Decision Makers", web_traffic_overtime, publishdates, "AI for Decision Makers") #AI for Decision Makers

p10 <- get_plot("Reproducibility in Cancer Informatics", web_traffic_overtime, publishdates, "Reproducibility in Cancer\nInformatics", useylab = TRUE) #Reproducibility in Cancer Informatics

p11 <- get_plot("Choosing Genomics Tools", web_traffic_overtime, publishdates, "Choosing Genomics\nTools") #Choosing Genomics Tools

p12 <- get_plot("Computing for Cancer Informatics", web_traffic_overtime, publishdates, "Computing for Cancer\nInformatics") #Computing for Cancer Informatics

p20 <- get_plot("Documentation and Usability", web_traffic_overtime, publishdates, "Documentation and\nUsability") #Documentation and Usability

p21 <- get_plot("Advanced Reproducibility", web_traffic_overtime, publishdates, "Advanced\nReproducibility") #Advanced Reproducibility

p22 <- get_plot("AI for Efficient Programming", web_traffic_overtime, publishdates, "AI for Efficient\nProgramming", usexlab = TRUE) #AI for Efficient Programming

p23 <- get_plot("GitHub Automation for Scientists", web_traffic_overtime, publishdates, "GitHub Automation for\nScientists") #GitHub Automation for Scientists
```




```{r}
layout <- 'ABCDE
           FGH##
           IJKL#'
overall_plot <- wrap_plots(A=p00, B=p01, C=p02, D=p03, E=p04, F=p10, G=p11, H=p12, I=p20, J=p21, K=p22, L=p23, design=layout) +plot_layout(ncol = 5, nrow =3, axes="collect", axis_titles = "collect", guides="collect") & theme(legend.position = 'bottom')

overall_plot
```

```{r}
ggsave(file.path("plots", "itn_course_traffic_plot_by_course_reordered.png"), plot = overall_plot, width = 14, height = 7)
```

### Number of learners by how long the course has been out

```{r}
itcr_course_data %>% 
  filter(!(website %in% c("ITN Website", "OTTR website", "metricminer.org"))) %>%
  ggplot(aes(x = coursera_launch, y = coursera_count, color = target_audience)) + 
  geom_point() + 
  theme_minimal() + 
  scale_color_manual(values=cbPalette)
```

```{r}
itcr_course_data$webAndEnrollmentTotals <- itcr_course_data %>%
  select(website_count, coursera_count, leanpub_count) %>% rowSums(na.rm = TRUE)

itcr_course_data %>% 
  filter(!(website %in% c("ITN Website", "OTTR website", "metricminer.org"))) %>%
  mutate(duration = today() - website_launch) %>%
  ggplot(aes(x = duration, y = webAndEnrollmentTotals, color = target_audience)) + 
  geom_point() + 
  theme(panel.grid = element_line("black", linewidth = 0.25), panel.background = element_blank(), panel.border = element_rect("black", fill=NA, linewidth=0.5)) +
  labs(x = "How long the course has been out",
       y = "Bookdown Views + Coursera & Leanpub Enrollments",
       color = "Target Audience") +
  scale_color_manual(values=cbPalette) + 
  ggrepel::geom_text_repel(aes(x = duration, y = webAndEnrollmentTotals, label = website), size = 4, vjust = - 1, na.rm = TRUE)

ggsave(file.path("plots", "itn_website_traffic_by_how_long_the_course_has_been_published.png"), width = 12, height = 8)
```
### Software engagement 

```{r}
web_traffic_overtime %>% 
  group_by(website, month_year) %>% 
  count() %>% 
  filter(!(website %in% c("ITN Website", "OTTR website", "metricminer.org"))) %>%
  ggplot(aes(y = n, x = month_year)) + 
  geom_bar(stat = "identity", fill = "pink") + 
  scale_x_date(date_labels = "%b %Y") + 
  ylab("Unique ITN software website visitors") +
  theme_minimal() +
  theme(axis.text.x=element_text(angle=60, hjust=1), 
        strip.text.x = element_text(size = 8)) 

ggsave(file.path("plots", "itn_software_website_traffic.png"), width = 10, height = 5)
```

#### Coursera

READY
```{r}
ggplot(itcr_course_data %>% filter(coursera_count > 0), aes(x = reorder(website, -coursera_count), y = coursera_count, fill = target_audience)) +
  geom_bar(stat = "identity", na.rm = TRUE) +
  theme_classic() +
  theme(axis.text.x = element_text(angle = 45, hjust=1)) +
  labs(x = NULL,
       y = "Coursera enrollments",
       fill = "Target Audience") +
  geom_text(aes(label = coursera_count), size = 3, vjust = - 1, na.rm = TRUE) +
  ylim(c(0, 1200)) + 
  scale_fill_manual(values=cbPalette)

ggsave(file.path("plots", "itn_coursera_enrollments.png"), width = 4, height = 2)
```

#### Leanpub

READY
```{r}
ggplot(itcr_course_data %>% filter(leanpub_count > 0) , aes(x = reorder(website, -leanpub_count), y = leanpub_count, fill = target_audience)) +
  geom_bar(stat = "identity", na.rm = TRUE) +
  theme_classic() +
  theme(axis.text.x = element_text(angle = 45, hjust=1)) +
  labs(x = NULL,
       y = "Leanpub enrollments",
       fill = "Target Audience") +
  geom_text(aes(label = leanpub_count), size = 3, vjust = - 1, na.rm = TRUE) +
  ylim(c(0, 40)) + 
  scale_fill_manual(values=cbPalette)

ggsave(file.path("plots", "itn_leanpub_enrollments.png"), width = 4, height = 2)
```

### Session Info 

```{r}
sessionInfo()
```
